// src/utils/ads/kayak/flights/types.ts
export type kayakInlineAdType = {
  rank: number;
  backgroundImageUrl: string;
  price: {
    price: number;
    currency: string;
    localizedPrice: string;
  };
  bookingButtonText: string;
  cpcEstimate: {
    price: number;
    currency: string;
    localizedPrice: string;
  };
  logoUrl: string;
  productType: string;
  companyName: string;
  site: string;
  deepLink: string;
  headline: string;
  trackUrl: string;
  impressionUrl: string;
  description: string;
};